+++ pkg/contexts/ocm/blobhandler/handlers/generic/npm/const.go
@@ -1,4 +1,4 @@
-package npmjs
+package npm
 
 import (
 	"github.com/open-component-model/ocm/pkg/contexts/credentials/cpi"
@@ -6,15 +6,15 @@ import (
 )
 
 const (
-	// CONSUMER_TYPE is the npmjs repository type.
-	CONSUMER_TYPE     = "Registry.npmjs.com"
+	// CONSUMER_TYPE is the npm repository type.
+	CONSUMER_TYPE     = "Registry.npm.com"
 	BLOB_HANDLER_NAME = "ocm/npmPackage"
 
-	// ATTR_USERNAME is the username attribute. Required for login at any npmjs registry.
+	// ATTR_USERNAME is the username attribute. Required for login at any npm registry.
 	ATTR_USERNAME = cpi.ATTR_USERNAME
-	// ATTR_PASSWORD is the password attribute. Required for login at any npmjs registry.
+	// ATTR_PASSWORD is the password attribute. Required for login at any npm registry.
 	ATTR_PASSWORD = cpi.ATTR_PASSWORD
-	// ATTR_EMAIL is the email attribute. Required for login at any npmjs registry.
+	// ATTR_EMAIL is the email attribute. Required for login at any npm registry.
 	ATTR_EMAIL = cpi.ATTR_EMAIL
 )
