+++ pkg/contexts/ocm/blobhandler/handlers/generic/npm/identity.go
@@ -1,4 +1,4 @@
-package npmjs
+package npm
 
 import (
 	"path"
