+++ pkg/contexts/ocm/blobhandler/handlers/init.go
@@ -1,7 +1,7 @@
 package handlers
 
 import (
-	_ "github.com/open-component-model/ocm/pkg/contexts/ocm/blobhandler/handlers/generic/npmjs"
+	_ "github.com/open-component-model/ocm/pkg/contexts/ocm/blobhandler/handlers/generic/npm"
 	_ "github.com/open-component-model/ocm/pkg/contexts/ocm/blobhandler/handlers/generic/ocirepo"
 	_ "github.com/open-component-model/ocm/pkg/contexts/ocm/blobhandler/handlers/oci/ocirepo"
 	_ "github.com/open-component-model/ocm/pkg/contexts/ocm/blobhandler/handlers/ocm/comparch"
